+++ SEAG_F/Core/Inc/Filters/CombFeedback.c
@@ -13,12 +13,12 @@ CombFeedback* initializeCombFeedback(uint32_t sampleRate, float delayS){
 
 }
 
-int16_t combFeedbackAppendSample(CombFeedback *filter, int16_t newSample){
+int32_t combFeedbackAppendSample(CombFeedback *filter, int32_t newSample){
 
     static const float b0 = 1;
     static const float am = 0.8;
 
-    int16_t output = b0*newSample - am*filter->outputSamples[filter->M-1]; // M - 1 see struct def for explanation
+    int32_t output = b0*newSample - am*filter->outputSamples[filter->M-1]; // M - 1 see struct def for explanation
 
     shiftBuffer(filter->outputSamples, filter->M, output);
