+++ SEAG_F/Core/Inc/Filters/CombFeedback.h
@@ -13,14 +13,14 @@ typedef struct{
     //y[n] = b_0x[n] - a_My[n-M]
 
     int M;
-    int16_t *outputSamples; // [0] is most recent output, [n-1] is n samples ago. so index is backwards
+    int32_t *outputSamples; // [0] is most recent output, [n-1] is n samples ago. so index is backwards
 
     //no need to save input samples
 
 }CombFeedback;
 
 CombFeedback* initializeCombFeedback(uint32_t sampleRate, float delayS);
-int16_t combFeedbackAppendSample(CombFeedback *filter, int16_t newSample);
+int32_t combFeedbackAppendSample(CombFeedback *filter, int32_t newSample);
 
 
 #endif //COMB_FEED_BACK_H
