+++ SEAG_F/Core/Inc/Filters/FilterUtility.c
@@ -1,7 +1,7 @@
 
 #include "FilterUtility.h"
 
-void shiftBuffer(int16_t *buffer, uint32_t bufferSize, int16_t newValue){
+void shiftBuffer(int32_t *buffer, uint32_t bufferSize, int32_t newValue){
 
     for(uint32_t i = bufferSize-1; i > 0; i--){
         buffer[i] = buffer[i-1];
