+++ SEAG_F/Core/Inc/Filters/FilterUtility.h
@@ -6,6 +6,6 @@
 #ifndef FILTER_UTLITY_H
 #define FILTER_UTILITY_H
 
-void shiftBuffer(int16_t *buffer, uint32_t bufferSize, int16_t newValue);
+void shiftBuffer(int32_t *buffer, uint32_t bufferSize, int32_t newValue);
 
 #endif //FILTER_UTILITY_H
